Combine docs from `master` and `master-loco` (initial)
This is an initial pass at bringing the docs for both legacy and loco styles
into the same branch.  Where the docs differ substantively (eg `README.md` and
`usage.md`) the files are now duplicated.

It's a little rough/ugly, but it's better than continuously porting other
things between the branches.